+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.0] - 2020-11-13
+
+### Added
+
+* Log level setting can be set for SDK and MQTT library separately
+* Certificates directory path is now configurable
+
 ## [0.1.1] - 2020-11-05
 
 This is an initial release